// src/todolist/dto/update-todolist.dto.ts
export class UpdateTodoListDto {
  readonly name?: string;
  readonly description?: string;
  readonly language?: 'th' | 'en';
  readonly status?: boolean;
  readonly date?: Date;
  readonly datecomplete?: Date;
  readonly assignee?: string;

  readonly subtasks?: {
    title: string;
    done?: boolean;
  }[];
}